modify: Model Order(long to double); feat: proceed order (ok: temp)

+++ app/src/main/java/com/example/foodorderingapp/Adapter/CartProductAdapter.java
@@ -133,7 +133,7 @@ public class CartProductAdapter extends RecyclerView.Adapter<CartProductAdapter.
                                                             public void onDataChange(@NonNull DataSnapshot snapshot) {
                                                                 Product product = snapshot.getValue(Product.class);
                                                                 int totalAmount = cart.getTotalAmount() - cartInfo.getAmount();
-                                                                double totalPrice = cart.getTotalPrice() - (long) (product.getProductPrice() * cartInfo.getAmount());
+                                                                double totalPrice = cart.getTotalPrice() - (product.getProductPrice() * cartInfo.getAmount());
 
                                                                 HashMap<String, Object> map = new HashMap<>();
                                                                 map.put("totalAmount", totalAmount);
